requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiohttp-retry==2.9.1
aiosignal==1.3.2
alembic==1.15.2
attrs==25.3.0
bcrypt==4.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
dnspython==2.7.0
email_validator==2.2.0
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.2.2
frozenlist==1.6.0
GeoAlchemy2==0.17.1
geographiclib==2.0
geopy==2.3.0
google-api-core==2.24.2
google-auth==2.39.0
google-cloud-speech==2.17.0
googleapis-common-protos==1.70.0
greenlet==3.2.1
grpcio==1.71.0
grpcio-status==1.62.3
gTTS==2.3.2
gunicorn==20.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
multidict==6.4.3
packaging==25.0
propcache==0.3.1
proto-plus==1.26.1
protobuf==4.25.7
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyJWT==2.10.1
python-decouple==3.8
python-dotenv==1.0.0
python-http-client==3.3.7
pytz==2025.2
razorpay==1.3.0
requests==2.32.3
rsa==4.9.1
sendgrid==6.11.0
setuptools==79.0.1
# Editable install with no version control (shrambandhu==0.1)
#-e /mnt/c/Users/gupta/Desktop/rrlsp/uoh/ShramBandhu
SQLAlchemy==2.0.40
starkbank-ecdsa==2.2.0
twilio==8.3.0
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
WTForms==3.2.1
yarl==1.20.0
